Rename classes and cleanup
