+++ src/main/java/samj/mario/editor/ChangeForegroundTileCommand.java
@@ -1,16 +1,16 @@
 package samj.mario.editor;
 
-import samj.mario.editor.data.TileDefinition;
+import samj.mario.editor.data.Tile;
 
 public class ChangeForegroundTileCommand implements EditorCommand {
 
     private final int x;
     private final int y;
-    private final TileDefinition newTile;
-    private final TileDefinition oldTile;
+    private final Tile newTile;
+    private final Tile oldTile;
     private final LevelEditor levelEditor;
 
-    public ChangeForegroundTileCommand(int x, int y, TileDefinition newTile, TileDefinition oldTile, LevelEditor levelEditor) {
+    public ChangeForegroundTileCommand(int x, int y, Tile newTile, Tile oldTile, LevelEditor levelEditor) {
         this.x = x;
         this.y = y;
         this.newTile = newTile;
